Move the projectile refresh to Bouncer. Bouncer now has an OnSecondUpdate(). Move projectile created tracking to GetDataHandler from Bouncer.
